Use session identifier instead of memoizing it
This patch uses the fact that `Zend\Expressive\Session\Session` now
implements `Zend\Expressive\Session\SessionIdentifierAwareInterface`.
When creating a session, it passes it the session identifier discovered
when introspecting cookies, if any. When persistence is invoked, it
pulls the identifier from the session.

This approach should solve concurrency issues, and makes it more clear
when a session cookie should be sent.