Can run tests several times with success
Changing session content on every test is necessary to ensure that
the session is considered changed and thus that a cookie is set in
the response.

Otherwise the session from previous test run would be picked up from
disk and seen as nto changed and response would not be altered thus
failing any subsequent test run.